@@ -1,16 +1,32 @@
 <?php
 
-/** @var \Illuminate\Database\Eloquent\Factory $factory */
+namespace Database\Factories;
 
 use App\Service;
 use Carbon\Carbon;
-use Faker\Generator as Faker;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-$factory->define(Service::class, function (Faker $faker) {
-    return [
-        'name' => 'Morning Worship Services (Online)',
-        'description' => '',
-        'service_date' => Carbon::now(),
-        'breeze_id' => '',
-    ];
-});
+class ServiceFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = Service::class;
+
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        return [
+            'name' => 'Morning Worship Services (Online)',
+            'description' => '',
+            'service_date' => Carbon::now(),
+            'breeze_id' => '',
+        ];
+    }
+}
